requirements.txt
aiogram
telethon
python-dotenv
apscheduler
aiohttp
